20241014 [IT] [JPA] Account - select, update, delete

+++ src/main/java/com/pro/spring/handler/UserHandler.java
@@ -9,11 +9,11 @@ public interface UserHandler {
 
     List<UserEntity> saveAll(List<UserEntity> users);
 
-    UserEntity findUserById(int id);
+    UserEntity findUserById(Long id);
 
     List<UserEntity> findAllUser();
 
-    UserEntity deleteUserById(int id);
+    UserEntity deleteUserById(Long id);
 
     List<UserEntity> deleteAllUser();
 }
